web/requirements.txt
Django==3.0.5
pytz==2020.1
django-environ==0.4.5
psycopg2-binary==2.8.5
django-filter==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
